requirements.txt
langchain
langchain-community
streamlit
python-dotenv
openai
faiss-cpu
groq
